Factor shuffle stuff into own file and struct
